Reintroduce previously removed changes
